server/pyproject.toml
[project]
name = "yanglegeyang-server"
version = "0.1.0"
description = "羊了个羊 - 后端服务"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.14.0",
    "pyjwt>=2.12.1",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
    "sqlalchemy[asyncio]>=2.0.49",
    "uvicorn[standard]>=0.46.0",
]
